+++ resources/kubelet_installation_binary.rb
@@ -22,12 +22,12 @@
 resource_name :kubelet_installation_binary
 
 property :kubelet_binary_url,      String, default: lazy { "https://storage.googleapis.com/kubernetes-release/release/v#{kubelet_version}/bin/linux/amd64/kubelet" }
-property :kubelet_binary_checksum, String, default: 'e350f6873e24741f6b995f266534bc4fdfeb0a594c4a37d688cf016d34ed074f'
-property :kubelet_version,         String, default: '1.12.1'
+property :kubelet_binary_checksum, String, default: 'c2af77f501c3164e80171903028d35c632366f53dec0c8419828d4e55d86146f'
+property :kubelet_version,         String, default: '1.17.0'
 property :kubelet_dir,             String, default: '/opt/bin'
 property :cni_tgz_url,             String, default: lazy { "https://github.com/containernetworking/plugins/releases/download/v#{cni_version}/cni-plugins-amd64-v#{cni_version}.tgz" }
-property :cni_tgz_checksum,        String, default: 'f04339a21b8edf76d415e7f17b620e63b8f37a76b2f706671587ab6464411f2d'
-property :cni_version,             String, default: '0.6.0'
+property :cni_tgz_checksum,        String, default: '3ca15c0a18ee830520cf3a95408be826cbd255a1535a38e0be9608b25ad8bf64'
+property :cni_version,             String, default: '0.7.5'
 property :cni_prefix_dir,          String, default: '/opt'
 
 action :create do