+++ test/integration/package/package_test.rb
@@ -24,7 +24,13 @@ end
 describe directory('/etc/kubernetes') do
   its('owner') { should eq 'root' }
   its('group') { should eq 'root' }
-  its('mode') { should eq 0600 }
+  its('mode') { should eq 0700 }
+end
+
+describe directory('/etc/kubernetes/manifests') do
+  its('owner') { should eq 'root' }
+  its('group') { should eq 'root' }
+  its('mode') { should eq 0755 }
 end
 
 describe file('/etc/kubernetes/config.yaml') do