+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 This file is used to list changes made in each version of the kubelet cookbook.
 
+# 0.3.0
+
+- Binary default kubelet version is now 1.17.0
+- Binary default cni version is now 0.7.5
+- Tested with kubelet 1.17.0, 1.16.4 and 1.15.7 on CentOS 7, Debian 10, Fedora 30 and Ubuntu 18.04
+- Creates manifests directory
+- remove dependency to systemd cookbook
+
 # 0.2.0
 
 - Added support of kubelet 1.12